requirements.txt
# ============================================================
# Web Framework & API Server
# ============================================================
fastapi==0.115.12
uvicorn==0.34.2
sse-starlette==3.0.2
jinja2


# ============================================================
# LangChain / LangGraph Ecosystem (Agents, Graphs, Memory)
# ============================================================
langchain==1.0.7
langchain-aws==1.0.0
langchain-openai==1.0.3
langchain-tavily
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43


# ============================================================
# LLM Clients & Embeddings
# ============================================================
openai==2.8.1
dashscope==1.23.3
tiktoken==0.12.0
numpy==2.2.5


# ============================================================
# Configuration, Validation & Serialization
# ============================================================
pydantic==2.11.7
pydantic-settings==2.9.1
python-dotenv==1.1.0
dataclasses-json==0.6.7


# ============================================================
# AWS SDK (AgentCore, Bedrock, Runtime Invocation)
# ============================================================
boto3==1.40.72


# ============================================================
# HTTP Clients & Async Utilities
# ============================================================
requests==2.32.3
aiohttp==3.11.18


# ============================================================
# Text, Document & File Processing
# ============================================================
beautifulsoup4==4.13.4
lxml==5.4.0
pypdf==5.5.0
reportlab==4.4.3
unstructured==0.17.2
unstructured-client==0.35.0
pillow==11.3.0


# ============================================================
# MCP / FastMCP (Tool Server & Client)
# ============================================================
mcp==1.13.1
fastmcp==2.12.2


# ============================================================
# Utilities (Logging, Monitoring, Progress)
# ============================================================
psutil==7.0.0
rich==14.1.0
tqdm==4.67.1
